main: Prefer buffer driver to polling driver
In c7fb43e0, we added a buffer driver for ambient light sensors, but we
added it to the list of drivers in a way that would favour polling
drivers. Since we started using IIO_SENSOR_PROXY_TYPE to tag devices,
and as those can support multiple drivers, we need to prefer the buffer
driver as it's the one that doesn't require constant monitoring.

This also fixes operations on some Dell laptops where the polling driver
is completely broken.

See https://gitlab.freedesktop.org/hadess/iio-sensor-proxy/-/merge_requests/352